Nginx configuration with ssl
Adds the nginx Dockerfile. The nginx_local.conf is a version that works
if you deploy locally with "local.reminisce.me" as a hostname.
In order for this to work you will also need the nginx/server_keys
folder to contain reminisce.me.key and reminisce.me.crt files. This
folder is ignored by git for obvious reasons.